chapp/requirements.txt
<<<<<<< HEAD
asgiref==3.5.0
Django==3.0
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
mysqlclient==2.1.0
phonenumberslite==8.12.46
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
xlwt==1.3.0
=======
asgiref==3.5.0
Django==3.0
django-phone-field==1.8.1
django-phonenumber-field==6.1.0
mysqlclient==2.1.0
phonenumberslite==8.12.46
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0
xlwt==1.3.0
>>>>>>> ef13bc491c34d343a46b83806cad5b77557231de
